Add files via upload
This project is made from image resources of cats and dogs which were gathered from Kaggle and checked for any corrupted images before going further with the code. Then using Keras further preprocessing of the entire dataset is done taking the image and batch size into consideration. Then using the matplotlyb function the basic classification of cats and dogs is completed. Then further classification has been done by making training models keeping the buffer size as 32 and rescalling the picture to 255 char, plotting model with keras by per pixel count rearangement of physical features extracted from every single picture and cscaling them out in order of what is making a cat and same for the dog. After the disection of the image into 128,256 and so on bits, the model is made so the AI can learn to identify from any picture of a cat n dog and categorize them accordingly.